Add start of company list
Added table and fillable properties to company model to allow user to
interact with the API. Created new route group for for company api.
Removed description from migration as it is not a requirement. Changed
resources so they return the correctly formated data to frontend.
Added functions to API actions to controller so they actually update
company model. Stable version of company list without relations.