Add new view for company/user relation
New view lists all current users and allows them to be added to a
company. This was a project requirement and levarages the previously
added Many to Many relation. New company controller action added to
store relations and one to show company with relations. BeforeEnter
method added the previous route to check if a company is loaded into
the store and re-direct back to company dashboard if not. This was
to help with testing and avoid errors on the route when reloading.